+++ mavros_msgs/CHANGELOG.rst
@@ -2,6 +2,12 @@
 Changelog for package mavros_msgs
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+1.4.0.1 (2020-10-21)
+--------------------
+* Add ekf_status plugin
+* Update EstimatorStatus to include EKF innovations
+* Contributors: Charlie Burge
+
 1.4.0 (2020-09-11)
 ------------------
 * play_tune: Assign tune format directly